+++ apps/web/app/(site)/tools/page.tsx
@@ -1,5 +1,6 @@
 import { Metadata } from "next";
-import { ToolsPageContent } from "../../components/ToolsPageContent";
+
+import { PageContent } from "./PageContent";
 
 export const metadata: Metadata = {
   title: "Online Tools | Free Browser-Based Utilities",
@@ -7,6 +8,4 @@ export const metadata: Metadata = {
     "Discover Cap's collection of free online tools for file conversion, video editing, and more. All tools run directly in your browser for maximum privacy.",
 };
 
-export default function ToolsPage() {
-  return <ToolsPageContent />;
-}
+export default PageContent;
